requirements.txt
openai
requests
azure-core
azure-search-documents
python-dotenv